[Refactor] url to image function 분리
