+++ url_to_image.py
@@ -1,0 +1,11 @@
+import cv2
+import requests
+import numpy as np
+
+
+def url_to_image(url):
+    response = requests.get(url)
+    image = np.asarray(bytearray(response.content), dtype="uint8")
+    image = cv2.imdecode(image, cv2.IMREAD_COLOR)
+
+    return image
